Auto-shortening of filename can be customized
max_filename_length: default 40. Sets maximum number of characters in a filename (plus path) before its shortened. Set to some large value (9999) to disable.

Example: 
~/.config/nvim/lua/plugins.lua =>
~/c/n/l/plugins.lua

narrow_window_size: default 84. If winwidth is less than or equal to this value, the filename will automatically be shorted. Set to 0 to disable.

Additionally, if the filename is longer than the winwidth - 30, the filename will be automatically shortened.